backend/requirements.txt
flask
google-genai
python-dotenv
Flask-Cors
moviepy
